First, create ApplicationCoordinator.
This is top-most Coordinator, that will take over much of the duties that would have been handled by AppDelegate.
Make it a subclass of NavigationCoordinator since our app is using UINC as main container controller.
Keep a weak reference to UIApplication, it can come in handy if we need to access it, or its delegate (AppDelegate).